moves files and simplify building
