+++ .gitignore
@@ -5,3 +5,4 @@
 obj/
 .*
 !/.gitignore
+*.dylib
