+++ src/common/helpers.c
