+++ src/common/helpers.h
