+++ src/osx/shake_private.h
@@ -1,21 +1,23 @@
 #ifndef _SHAKE_PRIVATE_H_
 #define _SHAKE_PRIVATE_H_
 
-#if defined(PLATFORM_LINUX)
-
-#include <dirent.h>
-#include <linux/input.h>
-#include "helpers.h"
-
-#define SHAKE_DIR_NODES		"/dev/input"
+#include <ForceFeedback/ForceFeedback.h>
+#include <IOKit/IOKitLib.h>
+#include "../common/helpers.h"
 
 #define BITS_PER_LONG		(sizeof(long) * 8)
 #define OFF(x)			((x)%BITS_PER_LONG)
 #define BIT(x)			(1UL<<OFF(x))
 #define LONG(x)			((x)/BITS_PER_LONG)
-#define test_bit(bit, array)	((array[LONG(bit)] >> OFF(bit)) & 1)
+#define test_bit(bit, array)	((array >> OFF(bit)) & 1)
 #define BITS_TO_LONGS(x) \
 	(((x) + 8 * sizeof (unsigned long) - 1) / (8 * sizeof (unsigned long)))
+
+typedef struct effectContainer
+{
+	int id;
+	FFEffectObjectReference effect;
+} effectContainer;
 
 typedef struct Shake_Device
 {
@@ -23,17 +25,13 @@ typedef struct Shake_Device
 	int id;
 	int capacity; /* Number of effects the device can play at the same time */
 	/* Platform dependent section */
-	int fd;
-	char *node;
-	unsigned long features[BITS_TO_LONGS(FF_CNT)];
-
+	io_service_t service;
+	FFDeviceObjectReference device;
+	listElement *effectList;
+	FFCAPABILITIES features;
 } Shake_Device;
 
 extern listElement *listHead;
 extern unsigned int numOfDevices;
-
-int nameFilter(const struct dirent *entry);
-
-#endif /* PLATFORM_LINUX */
 
 #endif /* _SHAKE_PRIVATE_H_ */
